refactor: HeaderBar Component core 폴더로 이동
